travel_project/requirements.txt
asgiref==3.8.1
certifi==2025.1.31
charset-normalizer==3.4.1
Django==4.2
djangorestframework==3.14.0
idna==3.10
mysqlclient==2.2.2
pytz==2025.1
requests==2.32.3
sqlparse==0.5.3
tzdata==2025.1
urllib3==2.3.0
